+++ jemallocator/tests/smoke.rs
@@ -6,9 +6,7 @@ static A: Jemalloc = Jemalloc;
 
 #[test]
 fn smoke() {
-    let mut a = Vec::new();
-    a.reserve(1);
-    a.push(3);
+    let _a: Vec<i32> = (0..3).collect();
 }
 
 /// https://github.com/rust-lang/rust/issues/45955